+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ensure display order of error list matches the display order of the form elements.
 - Implementation and accessibility testing of reCAPTCHA V3. [#570](https://github.com/cds-snc/platform-forms-client/issues/570)
 - Prevent submission of form for a delayed period of time to help prevent spam submissions
+- New API path `/api/notify-callback` to plug GC Notify callback feature
 
 ### Changed